+++ src/.internal/charts/stock/toolbar/IndicatorControl.ts
@@ -20,6 +20,7 @@ import { MovingAverageDeviation } from "../indicators/MovingAverageDeviation";
 import { MovingAverageEnvelope } from "../indicators/MovingAverageEnvelope";
 import { StandardDeviation } from "../indicators/StandardDeviation";
 import { TypicalPrice } from "../indicators/TypicalPrice";
+import { AverageTrueRange } from "../indicators/AverageTrueRange";
 import { MedianPrice } from "../indicators/MedianPrice";
 import { OnBalanceVolume } from "../indicators/OnBalanceVolume";
 import { PVT } from "../indicators/PVT";
@@ -45,7 +46,7 @@ import { StockIcons } from "./StockIcons";
 import * as $array from "../../../core/util/Array";
 import * as $type from "../../../core/util/Type";
 
-export type Indicators = "Acceleration Bands" | "Accumulation Distribution" | "Accumulative Swing Index" | "Aroon" | "Awesome Oscillator" | "Bollinger Bands" | "Bull Bear Power" | "Chaikin Money Flow" | "Chaikin Oscillator" | "Commodity Channel Index" | "Disparity Index" | "MACD" | "Momentum" | "Moving Average" | "Moving Average Cross" | "Moving Average Deviation" | "Moving Average Envelope" | "On Balance Volume" | "Price Volume Trend" | "Relative Strength Index" | "Standard Deviation" | "Stochastic Oscillator" | "Stochastic Momentum Index" | "Trix" | "Typical Price" | "Volume" | "Volume Profile" | "VWAP" | "Williams R" | "Median Price" | "ZigZag";
+export type Indicators = "Acceleration Bands" | "Accumulation Distribution" | "Accumulative Swing Index" | "Aroon" | "Average True Range" | "Awesome Oscillator" | "Bollinger Bands" | "Bull Bear Power" | "Chaikin Money Flow" | "Chaikin Oscillator" | "Commodity Channel Index" | "Disparity Index" | "MACD" | "Momentum" | "Moving Average" | "Moving Average Cross" | "Moving Average Deviation" | "Moving Average Envelope" | "On Balance Volume" | "Price Volume Trend" | "Relative Strength Index" | "Standard Deviation" | "Stochastic Oscillator" | "Stochastic Momentum Index" | "Trix" | "Typical Price" | "Volume" | "Volume Profile" | "VWAP" | "Williams R" | "Median Price" | "ZigZag";
 
 export interface IIndicator {
 	id: string;
@@ -360,6 +361,13 @@ export class IndicatorControl extends DropdownListControl {
 					legend: legend
 				});
 				break;
+			case "Average True Range":
+					indicator = AverageTrueRange.new(this.root, {
+						stockChart: stockChart,
+						stockSeries: stockSeries,
+						legend: legend
+					});
+					break;				
 			case "On Balance Volume":
 				indicator = OnBalanceVolume.new(this.root, {
 					stockChart: stockChart,